run-case!: Add a support for a :sidekick test option.
If test map has a :sidekick key value - execute a test[sidekick]
function in a background thread while the test is running and then wait for
it to complete at the end of the test.

This is helpful for instance when we want to create some additional load
on the cluster while a functional test is running.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>