cassandra/batch: add a sidekick task to a bridge-test-slow-net test
Added a cassandra-stress sidekick task for a bridge-test-slow-net.

A cassandra-stress command is as follows:

scylla-cassandra-stres write no-warmup duration=5m -rate threads=500 -mode native cql3 -node <n1 Node IP> -log file=<log file>

The idea behind the XXX-slow-net tests is to see how scylla behaves
when some network branches become slow and scylla's internal buffers
would begin to fill up.

Since the jepsen's test itself is not capable to stress scylla in
any way we need a cassandra-stress for this task.

We will run a regular test that will check the correctness as usual
and cassandra-stress in a background will provide the required load.

The logs of the cassandra-stress run will be written to the standard
logs location.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>